WIP: Add configuration file
Move loading of Markdown serialization options to a extension method for DocsConfiguration and use the same logic in command line runner and MSBuild tasks
